#426: Update to Iced 0.14.0

+++ src/gui/button.rs
@@ -15,7 +15,7 @@ use crate::{
     scan::game_filter,
 };
 
-const WIDTH: u16 = 125;
+const WIDTH: u32 = 125;
 
 fn template(content: Text, action: Option<Message>, style: Option<style::Button>) -> Element {
     Button::new(content.align_x(alignment::Horizontal::Center))
